+++ fabric-content-registries-v0-bridge/src/main/java/dev/loaderbridge/fabric/api/content/registry/FabricContentRegistriesBridgeMod.java
@@ -2,6 +2,7 @@ package dev.loaderbridge.fabric.api.content.registry;
 
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.TagsUpdatedEvent;
+import net.minecraftforge.event.brewing.BrewingRecipeRegisterEvent;
 import net.minecraftforge.event.furnace.FurnaceFuelBurnTimeEvent;
 import net.minecraftforge.fml.common.Mod;
 
@@ -9,12 +10,18 @@ import net.minecraftforge.fml.common.Mod;
 public final class FabricContentRegistriesBridgeMod {
     public FabricContentRegistriesBridgeMod() {
         MinecraftForge.EVENT_BUS.addListener(this::onFuel);
+        MinecraftForge.EVENT_BUS.addListener(this::onBrewingRecipes);
         MinecraftForge.EVENT_BUS.addListener(this::onTagsUpdated);
     }
 
     private void onFuel(FurnaceFuelBurnTimeEvent event) {
         BridgeContentRegistries.customFuel(event.getItemStack().getItem())
                 .ifPresent(event::setBurnTime);
+    }
+
+    private void onBrewingRecipes(BrewingRecipeRegisterEvent event) {
+        net.fabricmc.fabric.api.registry.FabricBrewingRecipeRegistryBuilder.BUILD
+                .invoker().build(event.getBuilder());
     }
 
     private void onTagsUpdated(TagsUpdatedEvent event) {